Correct the use of SPServices to run the C2WTS.
The SPServices account is not the right account to be running the claims to windows token service as.  The C2WTS account requires very special elevated permissions and adding the Services account to local admins will result in warnings displayed in the health analyzer.  Unfortunately I don't yet know how to automate configuring of local security policies as it is very low level from what I can google currently... however this is a much better step in the right direction of informing the user they need a specific account with very high risk permissions.